refactor: int => size_t

+++ include/rafi/trace/ICycle.h
@@ -38,19 +38,19 @@ public:
     virtual bool IsIoStateExist() const = 0;
     virtual bool IsNoteExist() const = 0;
 
-    virtual int GetOpEventCount() const = 0;
-    virtual int GetMemoryEventCount() const = 0;
-    virtual int GetTrapEventCount() const = 0;
+    virtual size_t GetOpEventCount() const = 0;
+    virtual size_t GetMemoryEventCount() const = 0;
+    virtual size_t GetTrapEventCount() const = 0;
 
     virtual uint64_t GetPc(bool isPhysical) const = 0;
-    virtual uint64_t GetIntReg(int index) const = 0;
-    virtual uint64_t GetFpReg(int index) const = 0;
+    virtual uint64_t GetIntReg(size_t index) const = 0;
+    virtual uint64_t GetFpReg(size_t index) const = 0;
 
     virtual void CopyNote(std::string* pOutNote) const = 0;
     virtual void CopyIo(NodeIo* pOutState) const = 0;
-    virtual void CopyOpEvent(NodeOpEvent* pOutEvent, int index) const = 0;
-    virtual void CopyMemoryEvent(NodeMemoryEvent* pOutEvent, int index) const = 0;
-    virtual void CopyTrapEvent(NodeTrapEvent* pOutEvent, int index) const = 0;
+    virtual void CopyOpEvent(NodeOpEvent* pOutEvent, size_t index) const = 0;
+    virtual void CopyMemoryEvent(NodeMemoryEvent* pOutEvent, size_t index) const = 0;
+    virtual void CopyTrapEvent(NodeTrapEvent* pOutEvent, size_t index) const = 0;
 };
 
 }}
