+++ src/librafi_trace/GdbCycle.cpp
@@ -119,17 +119,17 @@ bool GdbCycle::IsNoteExist() const
     return false;
 }
 
-int GdbCycle::GetOpEventCount() const
+size_t GdbCycle::GetOpEventCount() const
 {
     return 0;
 }
 
-int GdbCycle::GetMemoryEventCount() const
+size_t GdbCycle::GetMemoryEventCount() const
 {
     return 0;
 }
 
-int GdbCycle::GetTrapEventCount() const
+size_t GdbCycle::GetTrapEventCount() const
 {
     return 0;
 }
@@ -139,7 +139,7 @@ uint64_t GdbCycle::GetPc(bool isPhysical) const
     return isPhysical ? 0 : m_Pc;
 }
 
-uint64_t GdbCycle::GetIntReg(int index) const
+uint64_t GdbCycle::GetIntReg(size_t index) const
 {
     if (!(0 <= index && index < IntRegCount))
     {
@@ -149,7 +149,7 @@ uint64_t GdbCycle::GetIntReg(int index) const
     return m_IntRegs[index];
 }
 
-uint64_t GdbCycle::GetFpReg(int index) const
+uint64_t GdbCycle::GetFpReg(size_t index) const
 {
     (void)index;
     return 0;
@@ -161,21 +161,21 @@ void GdbCycle::CopyNote(std::string* pOutNote) const
     RAFI_NOT_IMPLEMENTED();
 }
 
-void GdbCycle::CopyOpEvent(NodeOpEvent* pOutEvent, int index) const
+void GdbCycle::CopyOpEvent(NodeOpEvent* pOutEvent, size_t index) const
 {
     (void)pOutEvent;
     (void)index;
     RAFI_NOT_IMPLEMENTED();
 }
 
-void GdbCycle::CopyMemoryEvent(NodeMemoryEvent* pOutEvent, int index) const
+void GdbCycle::CopyMemoryEvent(NodeMemoryEvent* pOutEvent, size_t index) const
 {
     (void)pOutEvent;
     (void)index;
     RAFI_NOT_IMPLEMENTED();
 }
 
-void GdbCycle::CopyTrapEvent(NodeTrapEvent* pOutEvent, int index) const
+void GdbCycle::CopyTrapEvent(NodeTrapEvent* pOutEvent, size_t index) const
 {
     (void)pOutEvent;
     (void)index;
